add: single share/copy animated button
