+++ src/components/common/animations/ShareOrCopyButton/data_share.json
